+++ dragon/ML_trainer/_dragon_sequence_trainer.py
@@ -594,6 +594,7 @@ class DragonSequenceTrainer(_BaseDragonTrainer):
                        target_names: Optional[list[str]] = None,
                        n_steps: int = 50,
                        max_sequence_bins: int = 40,
+                       wrap_text_width: int = 15,
                        verbose: int = 0):
         """
         Explains sequence model predictions using Captum's Integrated Gradients.
@@ -606,6 +607,7 @@ class DragonSequenceTrainer(_BaseDragonTrainer):
             target_names (list[str] | None): Optional target names for plotting. If None, attempts to extract from training dataset or model.
             n_steps (int): Number of steps for Integrated Gradients approximation.
             max_sequence_bins (int): Maximum number of bins for plot visualization if using long sequence length. (0 < max_sequence_bins <= 50)
+            wrap_text_width (int): Maximum character width for feature names in plots. Longer names will be wrapped.
             verbose (int): Verbosity level for Captum output.
         """
         dataset_to_use = explain_dataset if explain_dataset is not None else self.validation_dataset
@@ -668,6 +670,7 @@ class DragonSequenceTrainer(_BaseDragonTrainer):
             n_steps=n_steps,
             device=self.device,
             max_sequence_bins=max_sequence_bins,
+            wrap_text_width=wrap_text_width,
             verbose=verbose
         )